created fakes and a test for a fake skeleton owner
