+++ skeleton/src/main/java/com/dropbox/kaiken/skeleton/core/AppSkeletonScopedServices.kt
@@ -6,15 +6,9 @@ import com.dropbox.kaiken.skeleton.scoping.AppServices
 import com.dropbox.kaiken.skeleton.scoping.UserServices
 import com.dropbox.kaiken.skeleton.scoping.UserServicesProvider
 
-class AppSkeletonDelegate(
+class AppSkeletonScopedServices internal  constructor(
     override val component: SdkSpec
 ) : SkeletonScopedServices {
-
-    override val appServices: AppServices = component.getSkeletonConfig()
-        .scopedServicesFactory
-        .createAppServices(
-            component
-        )
 
     override val userServicesFactory = { appServices: AppServices, user: SkeletonUser ->
         component.getSkeletonConfig().scopedServicesFactory.createUserServices(
@@ -25,7 +19,11 @@ class AppSkeletonDelegate(
 
     override lateinit var userServicesProvider: UserServicesProvider
 
-    override fun provideAppServices(): AppServices = appServices
+    override fun provideAppServices(): AppServices = component.getSkeletonConfig()
+        .scopedServicesFactory
+        .createAppServices(
+            component
+        )
 
     override fun provideUserServicesOf(userId: String): UserServices? =
         userServicesProvider.provideUserServicesOf(userId)
