+++ skeleton/src/main/java/com/dropbox/kaiken/skeleton/core/KaikenAppServices.kt
@@ -2,8 +2,16 @@ package com.dropbox.kaiken.skeleton.core
 
 import com.dropbox.kaiken.skeleton.scoping.AppScope
 import com.dropbox.kaiken.skeleton.scoping.AppServices
+import com.dropbox.kaiken.skeleton.scoping.SingleIn
+import com.dropbox.kaiken.skeleton.scoping.Teardownable
 import com.dropbox.kaiken.skeleton.usermanagement.UserManager
+import com.squareup.anvil.annotations.ContributesBinding
 import com.squareup.anvil.annotations.ContributesTo
+import dagger.Module
+import dagger.Provides
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.MainScope
 
 /**
  * Holds the common dependencies for [AAppScope]
@@ -12,4 +20,24 @@ import com.squareup.anvil.annotations.ContributesTo
 @ContributesTo(AppScope::class)
 interface KaikenAppServices : AppServices {
     fun userManager(): UserManager
+}
+
+@ContributesTo(AppScope::class)
+interface CoroutineScopeBindings  {
+    fun coroutineScopes():CoroutineScopes
+}
+
+@ContributesBinding(AppScope::class)
+class RealCoroutineScopes:CoroutineScopes {
+        override val mainScope: CoroutineScope
+            get() = MainScope()
+        override val globalScope: CoroutineScope
+            get() = GlobalScope
+
+    }
+
+
+interface CoroutineScopes {
+    val mainScope:CoroutineScope
+    val globalScope:CoroutineScope
 }
