+++ skeleton/src/main/java/com/dropbox/kaiken/skeleton/core/SkeletonOwnerApplication.kt
@@ -8,5 +8,5 @@ abstract class SkeletonOwnerApplication : SkeletonOwner, Application() {
         appSkeleton = AppSkeletonInitializer.init(getSdkSpec()).appSkeletonDelegate
     }
 
-    override lateinit var appSkeleton: AppSkeletonDelegate
+    override lateinit var appSkeleton: AppSkeletonScopedServices
 }
