+++ skeleton/src/test/java/com/dropbox/kaiken/skeleton/scoping/fugazi/AppTeardownHelperProvider.kt
@@ -1,0 +1,10 @@
+package com.dropbox.kaiken.skeleton.scoping.fugazi
+
+import com.dropbox.kaiken.skeleton.scoping.AppTeardownHelper
+import com.dropbox.kaiken.skeleton.scoping.AuthOptionalActivityScope
+import com.squareup.anvil.annotations.ContributesTo
+
+@ContributesTo(AuthOptionalActivityScope::class)
+interface AppTeardownHelperProvider {
+    fun appTeardownHelper(): AppTeardownHelper
+}
