+++ skeleton/src/test/java/com/dropbox/kaiken/skeleton/scoping/fugazi/UserTeardownHelperProvider.kt
@@ -1,0 +1,11 @@
+package com.dropbox.kaiken.skeleton.scoping.fugazi
+
+import com.dropbox.kaiken.skeleton.scoping.AuthRequiredActivityScope
+import com.dropbox.kaiken.skeleton.scoping.UserTeardownHelper
+import com.squareup.anvil.annotations.ContributesTo
+
+@ContributesTo(AuthRequiredActivityScope::class)
+interface UserTeardownHelperProvider {
+    fun userTeardownHelper(): UserTeardownHelper
+//    fun appTeardownHelper():AppTeardownHelper
+}
